feat: 파일 기능 추가(테스트 예정)

+++ src/main/java/com/udemy/star_d_link/Dto/CourseReviewResponseDto.java
@@ -10,8 +10,8 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class CourseReviewResponseDto {
-    private Integer boardId;
-    private Integer userId;
+    private Long boardId;
+    private Long userId;
     private Integer likeCount;
     private String title;
     private String content;
